Add meta object

+++ src/operations/generate.js
@@ -1,23 +1,35 @@
 import fs from "fs";
-import { error, green, warning } from "../lib/chalk.js";
+import { red, green, yellow } from "../lib/chalk.js";
 
-export default function (dirname, name) {
+export default async function (dirname, name) {
     if (!name) {
-        console.log(warning("Please specify a name for the migration"));
+        console.log(yellow("Please specify a name for the migration"));
         return;
     }
 
     if (!fs.readdirSync(`${dirname}/migrations`)) {
-        console.log(error("Please run init first"));
+        console.log(red("Please run init first"));
         return;
     }
 
-    const timestamp = Date.now();
+    if (!fs.existsSync(`${dirname}/migrations/meta.json`)) {
+        console.log(red("meta.json is not found in migration folder"));
+        return;
+    }
+
     try {
+        const data = fs.readFileSync(`${dirname}/migrations/meta.json`, "utf8");
+        const meta = JSON.parse(data);
+
+        const timestamp = Date.now();
         const fileName = `${timestamp}_${name}.sql`;
+
         fs.writeFileSync(`${dirname}/migrations/${fileName}`, "");
-        console.log(green(`Migration ${fileName} created`));
+        meta.migrations.pending.push(fileName);
     } catch (error) {
         console.log(error);
+    } finally {
+        fs.writeFileSync(`${dirname}/migrations/meta.json`, JSON.stringify(meta));
+        console.log(green(`Migration ${fileName} created`));
     }
 }
